chore: add proposal headers api
